Created deskController Moved logic related to moving desk to desk controller
