+++ src/desk.js
@@ -1,19 +1,16 @@
-import { CODES } from './desk-constants';
 import { heightConverter } from './height-converter';
-import { deskHelpers, sleep } from './helpers';
-
-const BufferFrom = Buffer.from;
-
-const PREFLIGHT_TIME_DURATION = 200;
-const MOVE_TIME_DURATION = 500;
+import { deskHelpers } from './helpers';
 
 export class Desk {
   constructor (peripheral) {
     this.peripheral = peripheral;
 
-    this.moveToCharacteristic = null;
-    this.heightCharacteristic = null;
-    this.moveCharacteristic = null;
+    this.characteristics = {
+      move: null,
+      moveTo: null,
+      height: null
+    };
+
     this.moveToIntervalId = null;
   }
 
@@ -27,7 +24,7 @@ export class Desk {
   }
 
   setCharacteristic = (name, characteristic) => {
-    this[name] = characteristic;
+    this.characteristics[name] = characteristic;
   }
 
   setCustomPreflight = (preflightTimeDuration) => {
@@ -41,61 +38,13 @@ export class Desk {
 
   setCharacteristics = (characteristics) => {
     console.log('settings characteristics');
-    this.setCharacteristic('moveCharacteristic', deskHelpers.getMoveCharacteristic(characteristics));
-    this.setCharacteristic('heightCharacteristic', deskHelpers.getHeightCharacteristic(characteristics));
-    this.setCharacteristic('moveToCharacteristic', deskHelpers.getMoveToCharacteristic(characteristics));
-  }
-
-  moveUpAsync = async () => {
-    await this.moveCharacteristic.writeAsync(new BufferFrom(CODES.up, 'hex'), false);
-    // TODO: add check for speed to resolve?
-  }
-
-  moveDownAsync = async () => {
-    await this.moveCharacteristic.writeAsync(new BufferFrom(CODES.down, 'hex'), false);
-    // TODO: add check for speed to resolve?
-  }
-
-  preflightRequestAsync = async () => {
-    await this.moveCharacteristic.writeAsync(new BufferFrom(CODES.preflight, 'hex'), false);
-    await sleep(this.preflightTimeDuration || PREFLIGHT_TIME_DURATION);
-  }
-
-  moveToAsync = async (requestedHeight) => {
-    const moveLoop = await this.getMoveLoop(requestedHeight);
-    await this.preflightRequestAsync();
-    return moveLoop();
-  }
-
-  getMoveLoop = async (requestedHeight) => {
-    const shouldStopMoving = await this.getShouldStopMoving(requestedHeight);
-
-    const requestedHeightHex = heightConverter.toHexReversed(requestedHeight);
-
-    return () => new Promise((resolve, reject) => {
-      this.moveToIntervalId = setInterval(async () => {
-        const currentHeight = await this.getCurrentHeightAsync();
-        if (shouldStopMoving(currentHeight, requestedHeight)) {
-          clearInterval(this.moveToIntervalId);
-          resolve();
-        }
-        await this.moveAsync(requestedHeightHex);
-      }, MOVE_TIME_DURATION);
-    });
-  }
-
-  stopAsync = async () => {
-    clearInterval(this.moveToIntervalId);
-    await this.moveCharacteristic.writeAsync(new BufferFrom(CODES.stop, 'hex'), false);
-  }
-
-  moveAsync = async (requestedHeight) => {
-    const heightForTransmission = new BufferFrom(requestedHeight, 'hex');
-    await this.moveToCharacteristic.writeAsync(heightForTransmission, false);
+    this.setCharacteristic('move', deskHelpers.getMoveCharacteristic(characteristics));
+    this.setCharacteristic('height', deskHelpers.getHeightCharacteristic(characteristics));
+    this.setCharacteristic('moveTo', deskHelpers.getMoveToCharacteristic(characteristics));
   }
 
   getCurrentHeightBufferAsync = () => {
-    return this.heightCharacteristic.readAsync();
+    return this.characteristics.height.readAsync();
   }
 
   getCurrentHeightAsync = async () => {
@@ -110,12 +59,5 @@ export class Desk {
   disconnectAsync = async () => {
     await this.peripheral.disconnectAsync();
     console.log('\nDesk disconnected');
-  }
-
-  getShouldStopMoving = async (requestedHeight) => {
-    const isMovingUp = await this.getCurrentHeightAsync() > requestedHeight;
-    return isMovingUp
-      ? (current, requested) => current <= requested
-      : (current, requested) => current >= requested;
   }
 };
